Removed obsolete conversation files Load unidentified NPC data for some debugging

+++ UnityScripts/scripts/Magic/SpellEffects/SpellEffectAlly.cs
@@ -7,9 +7,9 @@ public class SpellEffectAlly : SpellEffect {
 		/// Backup the original state of the Npc
 		//public int OriginalState;
 		/// Backup the original attitude and goals of the npc
-		public int OriginalAttitude;
-		public int OriginalGtarg;
-		public int OriginalGoal;
+		public short OriginalAttitude;
+		public short OriginalGtarg;
+		public short OriginalGoal;
 
 		private NPC npc;
 		public bool WasActive;
