+++ UnityScripts/scripts/Magic/SpellEffects/SpellEffectConfusion.cs
@@ -8,9 +8,9 @@ public class SpellEffectConfusion : SpellEffect {
 		/// Backup the original state of the Npc
 		//public int OriginalState;
 		/// Backup the original attitude and goals of the npc
-		public int OriginalAttitude;
-		public int OriginalGtarg;
-		public int OriginalGoal;
+		public short OriginalAttitude;
+		public short OriginalGtarg;
+		public short OriginalGoal;
 
 		private NPC npc;
 		public bool WasActive;
